IQA-120: Move Subject Model to subject Subfolder
