+++ web-app/src/models/model_registration.py
@@ -1,4 +1,4 @@
-from .Subject import Subject  # noqa: F401
+from .subject import Subject  # noqa: F401
 from .user import User, UserRole, UserUserRole  # noqa: F401
 from .suggestquestion import QuizSuggestion, QuizSuggestionAnswer  # noqa: F401
 from .quiz import QuizQuestion, QuizAnswer  # noqa: F401
