+++ web-app/src/models/subject/Subject.py
@@ -1,4 +1,4 @@
-from .. import db
+from ... import db
 
 
 class Subject(db.Model):
