+++ web-app/src/services/SubjectService.py
@@ -2,7 +2,7 @@ from typing import List
 
 from .abstracts.AbcSubjectService import AbcSubjectService
 
-from ..models.Subject import Subject
+from ..models.subject.Subject import Subject
 from ..api.v1.dtos.SubjectDto import SubjectDto
 from ..modules.subjects.viewmodels.SubjectOverviewViewModel import SubjectOverviewViewModel
